requirements.txt
flask
PyPDF2
pdfminer.six
nltk
loguru
sentence-transformers
pinecone-client
openai
